refactor: fixed clippy warnings
