+++ src/core/mod.rs
@@ -1,8 +1,8 @@
-pub mod crawler;
+pub mod crawling;
 mod errors;
 pub mod retry;
 pub mod spider;
 
-pub use crawler::crawler::Crawler;
+pub use crawling::crawler::Crawler;
 pub use errors::{ScraperError, ScraperResult};
 pub use spider::{Spider, SpiderCallback};
